requirements.txt

python-telegram-bot==20.7
deepface==0.40.1
opencv-python-headless==4.9.0.80
